clean render, added anim behaviour
 - removed error return signature where there was no error case
 - implement simple animation behaviour for sprites

+++ src/render2d/util_types.zig
@@ -2,8 +2,8 @@ const zlm = @import("zlm");
 
 pub const TextureHandle = struct {
     id: c_int,
-    width: u32,
-    height: u32,
+    width: f32,
+    height: f32,
 };
 
 pub const UV = struct {
